requirements.txt
requests~=2.31.0
openai~=0.27.8
pandas~=1.5.3
beautifulsoup4~=4.11.1
fastapi~=0.96.0
python-dotenv~=1.0.0